Split shared wallet state from channel-specific shared state
