+++ packages/wallet/src/redux/states/shared.ts
@@ -1,8 +1,8 @@
 import { TransactionRequest } from 'ethers/providers';
 import { WalletEvent, DisplayAction } from 'magmo-wallet-client';
-import { Commitment } from 'fmg-core';
 import { WalletAction } from '../actions';
 import { WaitForChannel } from './opening';
+import { SharedChannelState } from './channels/shared';
 
 export interface OutboxState {
   displayOutbox?: DisplayAction;
@@ -30,46 +30,10 @@ export interface AdjudicatorKnown extends LoggedIn {
   adjudicator: string;
 }
 
-export interface SharedChannelState {
-  address: string;
-  privateKey: string;
-}
-
-export interface HasChannel<T extends SharedChannelState> extends AdjudicatorKnown {
+export const WALLET_INITIALIZED = 'WALLET.INITIALIZED';
+export interface Initialized<T extends SharedChannelState> extends AdjudicatorKnown {
+  stage: typeof WALLET_INITIALIZED;
   channelState: T;
-}
-
-export interface SignedCommitment {
-  commitment: Commitment;
-  signature: string;
-}
-
-export interface FirstCommitmentReceived extends SharedChannelState {
-  channelId: string;
-  libraryAddress: string;
-  ourIndex: number;
-  participants: [string, string];
-  channelNonce: number;
-  turnNum: number;
-  lastCommitment: SignedCommitment;
-  requestedTotalFunds: string;
-  requestedYourDeposit: string;
-}
-export interface ChannelOpen extends FirstCommitmentReceived {
-  penultimateCommitment: SignedCommitment;
-}
-export interface ChannelOpenAndTransactionExists extends ChannelOpen {
-  transactionHash: string;
-}
-export interface TransactionExists {
-  transactionHash: string;
-}
-export interface ChallengeExists extends ChannelOpen {
-  challengeExpiry?: number;
-}
-
-export interface UserAddressExists extends ChannelOpen {
-  userAddress: string;
 }
 
 // creators
@@ -85,61 +49,4 @@ export function loggedIn<T extends LoggedIn>(params: T): LoggedIn {
 export function adjudicatorKnown<T extends AdjudicatorKnown>(params: T): AdjudicatorKnown {
   const { networkId, adjudicator } = params;
   return { ...loggedIn(params), networkId, adjudicator };
-}
-
-export function baseChannelState<T extends SharedChannelState>(params: T): SharedChannelState {
-  const { address, privateKey } = params;
-  return {
-    address,
-    privateKey,
-  };
-}
-
-export function firstCommitmentReceived<T extends FirstCommitmentReceived>(
-  params: T,
-): FirstCommitmentReceived {
-  const {
-    channelId,
-    ourIndex,
-    participants,
-    channelNonce,
-    turnNum,
-    lastCommitment: lastPosition,
-    libraryAddress,
-    requestedTotalFunds,
-    requestedYourDeposit,
-  } = params;
-  return {
-    ...baseChannelState(params),
-    channelId,
-    ourIndex,
-    participants,
-    channelNonce,
-    turnNum,
-    lastCommitment: lastPosition,
-    libraryAddress,
-    requestedTotalFunds,
-    requestedYourDeposit,
-  };
-}
-
-export function channelOpen<T extends ChannelOpen>(params: T): ChannelOpen {
-  return {
-    ...firstCommitmentReceived(params),
-    penultimateCommitment: params.penultimateCommitment,
-  };
-}
-
-export function challengeExists<T extends ChallengeExists>(params: T): ChallengeExists {
-  return {
-    ...channelOpen(params),
-    challengeExpiry: params.challengeExpiry,
-  };
-}
-
-export function userAddressExists<T extends UserAddressExists>(params: T): UserAddressExists {
-  return {
-    ...challengeExists(params),
-    userAddress: params.userAddress,
-  };
 }
